requirements.txt
requests
beautifulsoup4
streamlit
numpy
Pillow
flask
keras
cvlib
tensorflow
opencv-python